% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{checkIfGainOrLoseAfterAltOneVariant}
\alias{checkIfGainOrLoseAfterAltOneVariant}
\title{Check if the number of uAUG, ployA signal, or Kozak in UTR changes after alt for one variant}
\usage{
checkIfGainOrLoseAfterAltOneVariant(variant, colNameBeforeAlt, colNameAfterAlt)
}
\arguments{
\item{variant}{variant info}

\item{colNameBeforeAlt}{the name of the column before alter, type of a string}

\item{colNameAfterAlt}{the name of the column after alter, type of a string}
}
\value{
whether the number gained, lost, or equal after alter. If colNameBeforeAlt or colNameAfterAlt does not exist in the variantsTable, return NA
}
\description{
\code{checkIfGainOrLoseAfterAlt} check if the number of uAUG, ployA signal, or Kozak in UTR changes after alt. This function only works to compare two "numeric" columns
}
\keyword{internal}
