% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{checkIfLostCodonAfterAlt}
\alias{checkIfLostCodonAfterAlt}
\title{Check if the codon lost after alt.}
\usage{
checkIfLostCodonAfterAlt(variantsTable, colNameAfterAlt, patterns)
}
\arguments{
\item{variantsTable}{variants table}

\item{colNameAfterAlt}{the name of the codon column after alter, type of a string. For example, name of column for the stop codon sequence after alt}

\item{patterns}{codon patterns, "ATG" for start codon, c("TAG", "TAA", "TGA") for stop codon}
}
\value{
A list containing TURE or FALSE or NA (if the codon sequence is NA). If colNameAfterAlt does not exist in the variantsTable, return NA
}
\description{
Check if the codon lost after alt.
}
\keyword{internal}
