% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{checkIfLostCodonAfterAltOneVariant}
\alias{checkIfLostCodonAfterAltOneVariant}
\title{Check if the codon lost after alt for one variant.}
\usage{
checkIfLostCodonAfterAltOneVariant(variant, colNameAfterAlt, patterns)
}
\arguments{
\item{variant}{variant info}

\item{colNameAfterAlt}{the name of the codon column after alter, type of a string. For example, name of column for the stop codon sequence after alt}

\item{patterns}{codon patterns, "ATG" for start codon, c("TAG", "TAA", "TGA") for stop codon}
}
\value{
A string concatenated TURE or FALSE or NA of all transcripts for that variant. If colNameAfterAlt is NA, return NA
}
\description{
Check if the codon lost after alt for one variant.
}
\keyword{internal}
