% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{checkIfRefMatchAndAltFragment}
\alias{checkIfRefMatchAndAltFragment}
\title{Check if the nucleotides in Ref column match the actual nucleotides in the fragment}
\usage{
checkIfRefMatchAndAltFragment(
  fragment,
  fragment_start,
  fragment_end,
  strand,
  ref,
  pos,
  alt,
  refAlwaysFromPlusStrand = TRUE
)
}
\arguments{
\item{fragment}{sequence within [fragment_start, fragment_end]}

\item{fragment_start}{start postion of the fragment}

\item{fragment_end}{end postion of the fragment}

\item{strand}{strand of the current sequence, 1 or -1}

\item{ref}{nucleotides in Ref column}

\item{pos}{variant postion in Pos column}

\item{alt}{nucleotides in Alt column}
}
\value{
ragment sequence after mutation or -1 if the nucleotides in Ref column do not match the actual nucleotides in the sequence
}
\description{
Check if the nucleotides in Ref column match the actual nucleotides in the fragment
}
\keyword{internal}
