% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{checkInputValid}
\alias{checkInputValid}
\title{Check if the variable table is a valid for UTR annotation}
\usage{
checkInputValid(variantsTable, sep = ",")
}
\arguments{
\item{variantsTable}{A data.table containing all variants}

\item{sep}{Delimiter used in variant file for concatenating multiple items in one column}
}
\value{
A data.table containing all valid variants
}
\description{
Check whether has 4 required columns, remove rows have NAs in the required columns, split rows if there're collapsed columns (items are separated by ,)
}
\keyword{internal}
