% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{countDNAPatternOneVariant}
\alias{countDNAPatternOneVariant}
\title{Count the number of DNA pattern for one variant}
\usage{
countDNAPatternOneVariant(variant, seqTable, dnaPattern, transcriptIdCol)
}
\arguments{
\item{variant}{variant info}

\item{seqTable}{sequences table containing two columns: ensembl_transcript_id and seq. seq is the sequence of the variantType region: e.g. 5' UTR sequence or 3' UTR sequence. The seqTable is output from function getSeqTable.}

\item{dnaPattern}{the string pattern that you want to count, for example "ATG"}

\item{transcriptIdCol}{the ensemble transcript id column}
}
\value{
number of DNA pattern of all transcripts for that variant.
}
\description{
\code{countDNAPattern} count the number of DNA pattern for one variant
}
\keyword{internal}
