% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{createEnsDbFromAH}
\alias{createEnsDbFromAH}
\title{Create a EnsDb database file from AnnotationHub}
\usage{
createEnsDbFromAH(species, ensemblVersion, dataDir)
}
\arguments{
\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}

\item{dataDir}{path to store database information}
}
\value{
save Db file into dataDir/species/ensemblVersion/
}
\description{
\code{createEnsDbFromAH} create a sqlite db for specified species and version
}
\keyword{internal}
