% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation_scores.R
\name{get_conservation_scores}
\alias{get_conservation_scores}
\title{Get conservation scores for variants}
\usage{
get_conservation_scores(variantsTable, conservationBwPath)
}
\arguments{
\item{variantsTable}{A variant table, positions are 1-based}

\item{conservationBwPath}{The path of the bigWig conservation track file download from UCSC Genome Browser}
}
\value{
A vector of strings, which are conservation scores of input variants. If a variant reference sequence has more than 1 nucleotide, the conservation scores of all its positions will be concatenated with ";".
}
\description{
Get conservation scores for variants
}
\keyword{internal}
