% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getCodon}
\alias{getCodon}
\title{Get the codon sequence for transcripts in transcriptIdColumn.}
\usage{
getCodon(variantsTable, seqTable, transcriptIdColumn, codonRegion)
}
\arguments{
\item{variantsTable}{variants table}

\item{seqTable}{cooresponding sequence table from Ensembl DB (either 5' UTR sequence or 3' UTR sequence). Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}

\item{codonRegion}{"startCodon" or "stopCodon"}
}
\value{
A list containing codon sequences (NA if the transcript id in that row is NA). If transcriptIdColumn does not exist in the variantsTable, return NA
}
\description{
\code{getCodon} Get the codon sequence for transcripts in transcriptIdColumn.
}
\keyword{internal}
