% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getCodonInAlt}
\alias{getCodonInAlt}
\title{Get the altered codon sequence for transcripts in transcriptIdColumn.}
\usage{
getCodonInAlt(
  variantsTable,
  transcriptsTable,
  transcriptIdColumn,
  codonPositionsColumn,
  codonSeqColumn
)
}
\arguments{
\item{variantsTable}{variants table}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}

\item{codonSeqColumn}{the codon sequence column output from getCodon}

\item{trascriptsTable}{transcripts structure table from ensembl with start and end columns for start and end coordinates of the target regions}

\item{codonPositionsColumnn}{the codon coordinates column output from getTranscriptIdsForCodonVariants}
}
\value{
A list containing codon sequences after alt (NA if the transcript id in that row is NA). If transcriptIdColumn, codonPositionsColumn or codonSeqColumn does not exist in the variantsTable, return NA
}
\description{
Get the altered codon sequence for transcripts in transcriptIdColumn.
}
\keyword{internal}
