% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getCodonInAltOneVariant}
\alias{getCodonInAltOneVariant}
\title{Get the altered codon sequence for transcripts in transcriptIdColumn for one variant.}
\usage{
getCodonInAltOneVariant(
  variant,
  transcriptsTable,
  transcriptIdColumn,
  codonPositionsColumn,
  codonSeqColumn
)
}
\arguments{
\item{variant}{variant info}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}

\item{codonSeqColumn}{the codon sequence column output from getCodon}

\item{trascriptsTable}{transcripts structure table from ensembl with start and end columns for start and end coordinates of the target regions}

\item{codonPositionsColumnn}{the codon coordinates column output from getTranscriptIdsForCodonVariants}
}
\value{
A string concatenated codon sequences after alt (NA if the transcript id in that row is NA).
}
\description{
Get the altered codon sequence for transcripts in transcriptIdColumn for one variant.
}
\keyword{internal}
