% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getCodonOneVariant}
\alias{getCodonOneVariant}
\title{Get the codon sequence of transcripts in transcriptIdColumn for one variant.}
\usage{
getCodonOneVariant(variant, seqTable, transcriptIdColumn, codonRegion)
}
\arguments{
\item{variant}{variant info}

\item{seqTable}{cooresponding sequence table from Ensembl DB (either 5' UTR sequence or 3' UTR sequence). Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}

\item{codonRegion}{"startCodon" or "stopCodon"}
}
\value{
A string concatenated codon sequences (NA if the transcript id in that row is NA) for one variant.
}
\description{
\code{getCodonOneVariant} Get the codon sequence of transcripts in transcriptIdColumn for one variant.
}
\keyword{internal}
