% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getEnsembl}
\alias{getEnsembl}
\title{Get Ensembl database object for the specified species}
\usage{
getEnsembl(species, ensemblVersion, dataDir)
}
\arguments{
\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl version number}

\item{dataDir}{Path to store database information}
}
\value{
Save BioMart object to dataDir
}
\description{
Get Ensembl database object for the specified species
}
\keyword{internal}
