% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{getFeatureFromInfo}
\alias{getFeatureFromInfo}
\title{Get feature information from INFO column of VCF file}
\usage{
getFeatureFromInfo(infoList)
}
\arguments{
\item{infoList}{INFO column of VCF variant file}
}
\value{
A data table of useful information extracted from INFO
}
\description{
Get feature information from INFO column of VCF file
}
\keyword{internal}
