% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getKozakPWM}
\alias{getKozakPWM}
\title{Get Kozak PWM}
\usage{
getKozakPWM(kozakStrenFile = NULL)
}
\arguments{
\item{kozakStrenFile}{Customed Kozak strength file}
}
\value{
Kozak PWM
}
\description{
Get Kozak PWM
}
\keyword{internal}
