% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getKozakScore}
\alias{getKozakScore}
\title{Calculate the Kozak score for each Kozak sequence, and concatenate the numbers with ",".}
\usage{
getKozakScore(variantsTable, kozakSeqs, kozakPWM)
}
\arguments{
\item{variantsTable}{variants table}

\item{kozakSeqs}{the column name of the kozak sequences}

\item{kozakPWM}{Kozak PWM output from consensusMatrix() and PWM()}
}
\value{
a list containing scores for the kozak sequences. Can be used to add new column to the variants table.
}
\description{
Calculate the Kozak score for each Kozak sequence, and concatenate the numbers with ",".
}
\keyword{internal}
