% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getKozakScoreOneVariant}
\alias{getKozakScoreOneVariant}
\title{Calculate the Kozak score for each Kozak sequence, and concatenate the numbers with ";" for one variant}
\usage{
getKozakScoreOneVariant(variant, kozakSeqs, kozakPWM)
}
\arguments{
\item{variant}{variant info}

\item{kozakSeqs}{the column name of the kozak sequences}

\item{kozakPWM}{Kozak PWM output from consensusMatrix() and PWM()}
}
\value{
a string concatenated scores for the kozak sequences.
}
\description{
Calculate the Kozak score for each Kozak sequence, and concatenate the numbers with "," for one variant
}
\keyword{internal}
