% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_annotation.R
\name{getLatestEnsemblVersion}
\alias{getLatestEnsemblVersion}
\title{Get the latest Ensembl version that available by querying with both biomRt and AnnotationHub}
\usage{
getLatestEnsemblVersion(species)
}
\arguments{
\item{species}{human or mouse}
}
\value{
Latest Ensembl version number
}
\description{
Get the latest Ensembl version that available by querying with both biomRt and AnnotationHub
}
\keyword{internal}
