% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getOrgName}
\alias{getOrgName}
\title{Convert species name to AnnotationHub acceptable species name}
\usage{
getOrgName(species)
}
\arguments{
\item{species}{human or mouse}
}
\value{
AnnotationHub acceptable species name
}
\description{
Convert species name to AnnotationHub acceptable species name
}
\keyword{internal}
