% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getRegionsForDiscretePos}
\alias{getRegionsForDiscretePos}
\title{Get codon region table used for getAltSequence}
\usage{
getRegionsForDiscretePos(positions)
}
\arguments{
\item{positions}{A vector of positions}
}
\value{
A data table of regions with start and end columns showing the input positions
}
\description{
Get codon region table used for getAltSequence
}
\keyword{internal}
