% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getSeqTable}
\alias{getSeqTable}
\title{Get UTR sequences of a list of transcripts}
\usage{
getSeqTable(
  transcriptIds,
  seqType,
  species,
  ensemblVersion,
  dataDir,
  downstream = NULL,
  upstream = NULL
)
}
\arguments{
\item{transcriptIds}{a list of Ensembl transcript ids}

\item{seqType}{Either '5utr' or '3utr'. 5utr' for 5' UTR sequence, '3utr' for 3' UTR sequences}

\item{species}{human or mouse}

\item{dataDir}{directory to store DB information}
}
\value{
sequences table containing columns: ensembl_transcript_id and seq
}
\description{
\code{getSeqTable} will query 5' UTR or 3' UTR sequences of a list of transcripts from Ensembl database
}
\keyword{internal}
