% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getTranscriptIds}
\alias{getTranscriptIds}
\title{Get protein coding transcript ids that overlap with each variant in the variant table}
\usage{
getTranscriptIds(variantsTable, species, ensemblVersion, dataDir)
}
\arguments{
\item{variantsTable}{variant table}

\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}

\item{dataDir}{path to store database information}
}
\value{
a list of protein coding transcript ids that overlap with each variant, if more than one transcripts overlap with a variant, they are concatenated with ";"
}
\description{
\code{getTranscriptIds} Get transcript ids that overlap with each variant in the variant table
}
\keyword{internal}
