% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getTranscriptIdsForOneUTRVariant}
\alias{getTranscriptIdsForOneUTRVariant}
\title{Find all the ensembl_transcript_id for a variant which fall in the specified region for one variant}
\usage{
getTranscriptIdsForOneUTRVariant(variant, transcriptRegions, start, end)
}
\arguments{
\item{variant}{one row of variants table}

\item{transcriptRegions}{transcripts structure table from ensembl}

\item{start}{column name of the start region in the transcriptRegions table}

\item{end}{column name of the end region in the transcriptRegions table}
}
\value{
A string representing a list of containing ensembl_transcript_id and collapsed with ;
}
\description{
\code{getTranscriptIdsForOneUTRVariant} find transcripts if a variant fall in their specified region
}
\keyword{internal}
