% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getTranscriptIdsForUTRVariants}
\alias{getTranscriptIdsForUTRVariants}
\title{Find the ensembl_transcript_id for utr variants genes, or cds}
\usage{
getTranscriptIdsForUTRVariants(variantsTable, transcriptRegions, checkRegion)
}
\arguments{
\item{variantsTable}{variants table}

\item{transcriptRegions}{transcripts structure table from Ensembl}

\item{checkRegion}{check if a transcript within a region: utr5 or utr3 or cds}
}
\value{
A list containing ensembl_transcript_id of each variant which its mutation Pos is within those transcripts checkRegion.
}
\description{
\code{getTranscriptIdsForUTRVariants} find transcripts if a variant fall in their specified region and return all transcripts id for each variant in the variant table
}
\keyword{internal}
