% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getTranscriptIdsOneVariant}
\alias{getTranscriptIdsOneVariant}
\title{Search the db to find protein coding transcript ids that overlap with a variant}
\usage{
getTranscriptIdsOneVariant(variant, species, ensemblVersion, dataDir)
}
\arguments{
\item{variant}{one variant}

\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}

\item{dataDir}{path to store database information}
}
\value{
protein coding transcript ids that overlap with the variant, if more than one transcripts overlap with the variant, they are concatenated with ";", if no transcript found, return NA.
}
\description{
\code{getTranscriptIdsOneVariant} Get transcript ids that overlap with one variant
}
\keyword{internal}
