% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{getTrasncriptsRegions}
\alias{getTrasncriptsRegions}
\title{Get information on transcript regions, UTRs regions, coding regions, chromosome name, and strand from Ensembl database}
\usage{
getTrasncriptsRegions(transcriptIds, species, ensemblVersion, dataDir)
}
\arguments{
\item{transcriptIds}{Ensembl transcript ids}

\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}

\item{dataDir}{directory to store DB information}
}
\value{
transcript structure table containing "ensembl_gene_id", "ensembl_transcript_id", "transcript_start", "transcript_end", "5_utr_start", "5_utr_end", "3_utr_start", "3_utr_end", "genomic_coding_start", "genomic_coding_end","gene_biotype", "strand"
}
\description{
\code{getTrasncriptsRegions} query "ensembl_gene_id", "ensembl_transcript_id", "transcript_start", "transcript_end", "5_utr_start", "5_utr_end", "3_utr_start", "3_utr_end", "genomic_coding_start", "genomic_coding_end", "gene_biotype", "strand" from Ensembl database
}
\keyword{internal}
