% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getTSSKozak}
\alias{getTSSKozak}
\title{Get the Kozak sequence in TSS region for transcripts in transcriptIdColumn.}
\usage{
getTSSKozak(variantsTable, utr5SeqTable, codingSeqTable, transcriptIdColumn)
}
\arguments{
\item{variantsTable}{variants table}

\item{utr5SeqTable}{5' UTR sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{codingSeqTable}{coding sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}
}
\value{
A list containing TSS Kozak sequences (NA if the transcript id in that row is NA). If transcriptIdColumn does not exist in the variantsTable, return NA
}
\description{
Get the Kozak sequence in TSS region for transcripts in transcriptIdColumn.
}
\keyword{internal}
