% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{getTSSKozakOneVariant}
\alias{getTSSKozakOneVariant}
\title{Get the Kozak sequence in TSS region for transcripts in transcriptIdColumn for one variant.}
\usage{
getTSSKozakOneVariant(
  variant,
  utr5SeqTable,
  codingSeqTable,
  transcriptIdColumn
)
}
\arguments{
\item{utr5SeqTable}{5' UTR sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{codingSeqTable}{coding sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column for the codon region}

\item{variantsTable}{variants table}
}
\value{
A string concatenated TSS Kozak sequences (NA if the transcript id in that row is NA)
}
\description{
Get the Kozak sequence in TSS region for transcripts in transcriptIdColumn for one variant.
}
\keyword{internal}
