% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr_analysis.R
\name{init_backend}
\alias{init_backend}
\title{Create parallel backend with user specified number of CPUs}
\usage{
init_backend(cores)
}
\arguments{
\item{cores}{number of CPUs}
}
\value{
Register the parallel backend with user specified number of cores
}
\description{
Create parallel backend with user specified number of CPUs
}
\keyword{internal}
