% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr5_mrl_prediction.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Reverse the scaling of the predicted MRL}
\usage{
inverse_transform(mrl)
}
\arguments{
\item{mrl}{MRL prediction}
}
\value{
inverse transformed MRL prediction
}
\description{
Reverse the scaling of the predicted MRL
}
\keyword{internal}
