% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr5_mrl_prediction.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{Apply one hot encode to the 100nt 5' UTR sequence}
\usage{
one_hot_encode(utr5_seq)
}
\arguments{
\item{utr5_seq}{100nt DNA sequence}
}
\value{
an array with dimension as c(1, 100, 4)
}
\description{
\code{one_hot_encode} will encode 5' UTR sequence of 100nt
}
\keyword{internal}
