% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr5_mrl_prediction.R
\name{predictMRL}
\alias{predictMRL}
\title{Get the 100nt 5' UTR sequence upstream of start codon for transcripts in transcriptIdColumn and predict their MRL}
\usage{
predictMRL(variantsTable, seqTable, transcriptIdColumn)
}
\arguments{
\item{variantsTable}{variants table}

\item{seqTable}{corresponding 5' UTR sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column}
}
\value{
A list predicted MRL (NA if the transcript id in that row is NA). If transcriptIdColumn does not exist in the variantsTable, return NA
}
\description{
\code{predictMRL} Get the 100nt 5' UTR sequence upstream of start codon for transcripts in transcriptIdColumn and predict their MRL
}
\keyword{internal}
