% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr5_mrl_prediction.R
\name{predictMRLInAltOneVariant}
\alias{predictMRLInAltOneVariant}
\title{Predict altered MRL for one variant}
\usage{
predictMRLInAltOneVariant(
  variant,
  seqTable,
  transcriptsTable,
  transcriptIdColumn,
  model
)
}
\arguments{
\item{variant}{A variant}

\item{seqTable}{corresponding 5' UTR sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{transcriptsTable}{transcripts structure table from ensembl with start and end columns for start and end coordinates of the 5' UTR}

\item{transcriptIdColumn}{the name of the transcript id column}

\item{model}{MRL prediction model}
}
\value{
MRL prediction of altered 5' UTR for the input variant
}
\description{
Predict altered MRL for one variant
}
\keyword{internal}
