% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utr5_mrl_prediction.R
\name{predictMRLOneVariant}
\alias{predictMRLOneVariant}
\title{Predict MRL for one variant}
\usage{
predictMRLOneVariant(variant, seqTable, transcriptIdColumn, model)
}
\arguments{
\item{variant}{A variant}

\item{seqTable}{corresponding 5' UTR sequence table from Ensembl DB. Have two columns: ensembl_transcript_id and seq}

\item{transcriptIdColumn}{the name of the transcript id column}

\item{model}{MRL prediction model}
}
\value{
MRL prediction of reference 5' UTR for the input variant
}
\description{
Predict MRL for one variant
}
\keyword{internal}
