% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{queryEnsemblInfo}
\alias{queryEnsemblInfo}
\title{query transcrips regions, utr sequences, and coding sequences from Ensembl database}
\usage{
queryEnsemblInfo(
  transcriptIds,
  species,
  ensemblVersion,
  dataDir,
  verbose = FALSE
)
}
\arguments{
\item{transcriptIds}{Ensembl transcript ids}

\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}

\item{dataDir}{path to store database information}

\item{verbose}{Whether print diagnostic messages. The default is FALSE.}
}
\value{
save all Db objects into dataDir
}
\description{
\code{queryEnsemblInfo} query transcrips regions, utr sequences, and coding sequences from Ensembl database and save rds into sepecified directory
}
\keyword{internal}
