% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readVariantData}
\alias{readVariantData}
\title{Read variant file in CSV format}
\usage{
readVariantData(variantFile, format = "csv")
}
\arguments{
\item{variantFile}{A varian file in CSV or VCF format}

\item{format}{File format, csv or vcf. The default file format is csv}
}
\value{
A data.table containing all variants if variant file is valid
}
\description{
\code{readVariantData} Read variant file in CSV / VCF format, check if it contains required columns for UTR annotation, split rows if there're collapsed columns (items are separated by ,), and remove rows containing NAs in Chr, Pos, Ref or Alt
}
\keyword{internal}
