% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{readVCFData}
\alias{readVCFData}
\title{Read variants from a VCF formatted file}
\usage{
readVCFData(variantFile)
}
\arguments{
\item{variantFile}{A variant file in VCF format}
}
\value{
A variant table
}
\description{
Read variants from a VCF formatted file
}
\keyword{internal}
