% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{splitRowsIfMultiFeature}
\alias{splitRowsIfMultiFeature}
\title{Split rows if the feature column(s) contains multiple items, separated by sep}
\usage{
splitRowsIfMultiFeature(variantsTable, columnNames, sep = ",")
}
\arguments{
\item{variantsTable}{variants table (data.table)}

\item{columnNames}{a string or a vector of strings that specifies columns need to be checked and splited}

\item{sep}{separator that is used to seperate items in the column(s)}
}
\value{
A new data.table that contains one item in the specified column(s) in each row
}
\description{
Split rows if the feature column(s) contains multiple items, separated by sep
}
\details{
splitRowsIfMultiFeature(variants, "Alt"): check column "Alt" to see if any rows have multiple Alt sequences like ATG,TC; if so, split these rows so that one row only has one Alt sequences.
}
\keyword{internal}
