% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_db.R
\name{validateTranscripts}
\alias{validateTranscripts}
\title{Checking whether a transcript is valid}
\usage{
validateTranscripts(transcriptIds, species, ensemblVersion, dataDir = "data")
}
\arguments{
\item{transcriptIds}{a list of Ensembl transcript ids}

\item{species}{human or mouse}

\item{ensemblVersion}{Ensembl database version}
}
\value{
transcript_regions with additional flags: protein_coding, valid_startcodon, valid_stopcodon
}
\description{
\code{validateTranscripts} checks 1) whether a transcript has a peptide sequence, 2) whether the peptide sequence ends with "*", 3) whether the peptide starts with "M"
}
\keyword{internal}
